tests build with msvc now

+++ Utils/TestArgumentParsing.c
@@ -126,7 +126,7 @@ TestFunction(TestArgumentParsing1){
         const char* argv[1] = {NULL};
         Args args= {0, argv};
         enum ArgParseError e = parse_args(&parser, &args, ARGPARSE_FLAGS_NONE);
-        TestExpectEquals(e, ARGPARSE_NO_ERROR);
+        TestExpectEquals((int)e, ARGPARSE_NO_ERROR);
         clear_parser(&parser);
         memset(&h, 0, sizeof(h));
     }
@@ -134,7 +134,7 @@ TestFunction(TestArgumentParsing1){
         const char* argv[] = {"--f"};
         Args args= {arrlen(argv), argv};
         enum ArgParseError e = parse_args(&parser, &args, ARGPARSE_FLAGS_NONE);
-        TestExpectEquals(e, ARGPARSE_NO_ERROR);
+        TestExpectEquals((int)e, ARGPARSE_NO_ERROR);
         TestExpectTrue(h.flag);
         clear_parser(&parser);
         memset(&h, 0, sizeof(h));
@@ -143,7 +143,7 @@ TestFunction(TestArgumentParsing1){
         const char* argv[] = {"--f", "--f"};
         Args args= {arrlen(argv), argv};
         enum ArgParseError e = parse_args(&parser, &args, ARGPARSE_FLAGS_NONE);
-        TestExpectEquals(e, ARGPARSE_DUPLICATE_KWARG);
+        TestExpectEquals((int)e, ARGPARSE_DUPLICATE_KWARG);
         TestExpectTrue(h.flag);
         clear_parser(&parser);
         memset(&h, 0, sizeof(h));
@@ -172,7 +172,7 @@ TestFunction(TestArgumentParsing2){
         const char* argv[] = {"--f", "lol"};
         Args args = {arrlen(argv), argv};
         enum ArgParseError e = parse_args(&parser, &args, ARGPARSE_FLAGS_NONE);
-        TestExpectEquals(e, ARGPARSE_NO_ERROR);
+        TestExpectEquals((int)e, ARGPARSE_NO_ERROR);
         clear_parser(&parser);
         memset(&f, 0, sizeof(f));
     }
@@ -180,7 +180,7 @@ TestFunction(TestArgumentParsing2){
         const char* argv[] = {"--f", "-g", "lol"};
         Args args = {arrlen(argv), argv};
         enum ArgParseError e = parse_args(&parser, &args, ARGPARSE_FLAGS_NONE);
-        TestExpectEquals(e, ARGPARSE_UNKNOWN_KWARG);
+        TestExpectEquals((int)e, ARGPARSE_UNKNOWN_KWARG);
         clear_parser(&parser);
         memset(&f, 0, sizeof(f));
     }
@@ -217,7 +217,7 @@ TestFunction(TestArgumentParsing3){
     };
     Args args = {arrlen(argv), argv};
     enum ArgParseError e = parse_args(&argparser, &args, ARGPARSE_FLAGS_NONE);
-    TestExpectEquals(e, 0);
+    TestExpectEquals((int)e, 0);
     TestExpectEquals(foo, 3.0f);
     TestExpectEquals(bar, -1e12);
     TESTEND();
@@ -255,7 +255,7 @@ TestFunction(TestArgumentParsing4){
         const char* argv[] = {"a1", "-b", "b1", "c"};
         Args args = {arrlen(argv), argv};
         enum ArgParseError e = parse_args(&argparser, &args, ARGPARSE_FLAGS_NONE);
-        TestExpectEquals(e, ARGPARSE_EXCESS_ARGS);
+        TestExpectEquals((int)e, ARGPARSE_EXCESS_ARGS);
         clear_parser(&argparser);
         memset(a, 0, sizeof(a));
         memset(b, 0, sizeof(b));
@@ -264,7 +264,7 @@ TestFunction(TestArgumentParsing4){
         const char* argv[] = {"a1", "a2", "-b", "b1"};
         Args args = {arrlen(argv), argv};
         enum ArgParseError e = parse_args(&argparser, &args, ARGPARSE_FLAGS_NONE);
-        TestExpectEquals(e, 0);
+        TestExpectEquals((int)e, 0);
         clear_parser(&argparser);
         memset(a, 0, sizeof(a));
         memset(b, 0, sizeof(b));
@@ -273,7 +273,7 @@ TestFunction(TestArgumentParsing4){
         const char* argv[] = {"-b", "b1", "a1", "a2"};
         Args args = {arrlen(argv), argv};
         enum ArgParseError e = parse_args(&argparser, &args, ARGPARSE_FLAGS_NONE);
-        TestExpectEquals(e, 0);
+        TestExpectEquals((int)e, 0);
         clear_parser(&argparser);
         memset(a, 0, sizeof(a));
         memset(b, 0, sizeof(b));
@@ -333,7 +333,7 @@ TestFunction(TestParseUserDefined){
         const char* argv[] = {"asd"};
         Args args = {arrlen(argv), argv};
         enum ArgParseError e = parse_args(&argparser, &args, ARGPARSE_FLAGS_NONE);
-        TestExpectEquals(e, ARGPARSE_CONVERSION_ERROR);
+        TestExpectEquals((int)e, ARGPARSE_CONVERSION_ERROR);
         clear_parser(&argparser);
         memset(&p, 0, sizeof(p));
     }
@@ -341,7 +341,7 @@ TestFunction(TestParseUserDefined){
         const char* argv[] = {"-1,3"};
         Args args = {arrlen(argv), argv};
         enum ArgParseError e = parse_args(&argparser, &args, ARGPARSE_FLAGS_NONE);
-        TestExpectEquals(e, 0);
+        TestExpectEquals((int)e, 0);
         TestExpectEquals(p.x, -1);
         TestExpectEquals(p.y, 3);
         clear_parser(&argparser);
@@ -351,7 +351,7 @@ TestFunction(TestParseUserDefined){
         const char* argv[] = {"4,6"};
         Args args = {arrlen(argv), argv};
         enum ArgParseError e = parse_args(&argparser, &args, ARGPARSE_FLAGS_NONE);
-        TestExpectEquals(e, 0);
+        TestExpectEquals((int)e, 0);
         TestExpectEquals(p.x, 4);
         TestExpectEquals(p.y, 6);
         clear_parser(&argparser);
@@ -396,7 +396,7 @@ TestFunction(TestParseEnum){
         const char* argv[] = {"asd"};
         Args args = {arrlen(argv), argv};
         enum ArgParseError e = parse_args(&argparser, &args, ARGPARSE_FLAGS_NONE);
-        TestExpectEquals(e, ARGPARSE_CONVERSION_ERROR);
+        TestExpectEquals((int)e, ARGPARSE_CONVERSION_ERROR);
         clear_parser(&argparser);
     }
     {
@@ -404,8 +404,8 @@ TestFunction(TestParseEnum){
         const char* argv[] = {"no-foo-bar"};
         Args args = {arrlen(argv), argv};
         enum ArgParseError e = parse_args(&argparser, &args, ARGPARSE_FLAGS_NONE);
-        TestExpectEquals(e, 0);
-        TestExpectEquals(fb, NOFOOBAR);
+        TestExpectEquals((int)e, 0);
+        TestExpectEquals((int)fb, NOFOOBAR);
         clear_parser(&argparser);
     }
     {
@@ -413,8 +413,8 @@ TestFunction(TestParseEnum){
         const char* argv[] = {"foo"};
         Args args = {arrlen(argv), argv};
         enum ArgParseError e = parse_args(&argparser, &args, ARGPARSE_FLAGS_NONE);
-        TestExpectEquals(e, 0);
-        TestExpectEquals(fb, FOO);
+        TestExpectEquals((int)e, 0);
+        TestExpectEquals((int)fb, FOO);
         clear_parser(&argparser);
     }
     TESTEND();
@@ -447,7 +447,7 @@ TestFunction(TestParseHex){
     #define FailHexTest(hexstr, error_code) do{\
         char argstring[] = hexstr;\
         Uint64Result e = parse_hex(argstring, sizeof(argstring)-1);\
-        TestExpectEquals(e.errored, error_code);\
+        TestExpectEquals((int)e.errored, error_code);\
     }while(0)
 
     FailHexTest("0 xff",                PARSENUMBER_INVALID_CHARACTER);
@@ -628,7 +628,7 @@ TestFunction(TestBitFlags){
         const char* argv[] = {"--foo"};
         Args args = {arrlen(argv), argv};
         enum ArgParseError e = parse_args(&argparser, &args, ARGPARSE_FLAGS_NONE);
-        TestExpectEquals(e, 0);
+        TestExpectEquals((int)e, 0);
         TestExpectEquals(flags, 1);
     }
     {
@@ -637,7 +637,7 @@ TestFunction(TestBitFlags){
         const char* argv[] = {"--dango"};
         Args args = {arrlen(argv), argv};
         enum ArgParseError e = parse_args(&argparser, &args, ARGPARSE_FLAGS_NONE);
-        TestExpectEquals(e, 0);
+        TestExpectEquals((int)e, 0);
         TestExpectEquals(flags, 4);
     }
     {
@@ -646,7 +646,7 @@ TestFunction(TestBitFlags){
         const char* argv[] = {"--bar", "--dango"};
         Args args = {arrlen(argv), argv};
         enum ArgParseError e = parse_args(&argparser, &args, ARGPARSE_FLAGS_NONE);
-        TestExpectEquals(e, 0);
+        TestExpectEquals((int)e, 0);
         TestExpectEquals(flags, 6);
     }
     TESTEND();
@@ -700,7 +700,7 @@ TestFunction(TestAppender){
         const char* argv[] = {"asd"};
         Args args = {arrlen(argv), argv};
         enum ArgParseError e = parse_args(&argparser, &args, ARGPARSE_FLAGS_NONE);
-        TestExpectEquals(e, ARGPARSE_CONVERSION_ERROR);
+        TestExpectEquals((int)e, ARGPARSE_CONVERSION_ERROR);
         TestExpectEquals(shorts.count, 0);
         Marray_cleanup(short)(&shorts, ctx.a);
         clear_parser(&argparser);
@@ -709,7 +709,7 @@ TestFunction(TestAppender){
         const char* argv[] = {"-1"};
         Args args = {arrlen(argv), argv};
         enum ArgParseError e = parse_args(&argparser, &args, ARGPARSE_FLAGS_NONE);
-        TestExpectEquals(e, ARGPARSE_INSUFFICIENT_ARGS);
+        TestExpectEquals((int)e, ARGPARSE_INSUFFICIENT_ARGS);
         TestAssertEquals(shorts.count, 1);
         TestExpectEquals(shorts.data[0], -1);
         Marray_cleanup(short)(&shorts, ctx.a);
@@ -719,7 +719,7 @@ TestFunction(TestAppender){
         const char* argv[] = {"4", "6", "8", "10", "12"};
         Args args = {arrlen(argv), argv};
         enum ArgParseError e = parse_args(&argparser, &args, ARGPARSE_FLAGS_NONE);
-        TestExpectEquals(e, 0);
+        TestExpectEquals((int)e, 0);
         TestAssertEquals(shorts.count, 5);
         TestExpectEquals(shorts.data[0], 4);
         TestExpectEquals(shorts.data[1], 6);
@@ -733,7 +733,7 @@ TestFunction(TestAppender){
         const char* argv[] = {"262144"};
         Args args = {arrlen(argv), argv};
         enum ArgParseError e = parse_args(&argparser, &args, ARGPARSE_FLAGS_NONE);
-        TestExpectEquals(e, ARGPARSE_CONVERSION_ERROR);
+        TestExpectEquals((int)e, ARGPARSE_CONVERSION_ERROR);
         TestExpectEquals(shorts.count, 0);
         Marray_cleanup(short)(&shorts, ctx.a);
         clear_parser(&argparser);
